requirements.txt

ipython
boltons
pytest
pytest-env
factory_boy
spooky
sqlalchemy
jupyter
blessings
wordfreq
numpy
scandir
psutil
mpi4py
anyconfig
PyYAML
invoke
matplotlib
flake8
dataset
lxml
ujson
